Relocate settings subcomponents out of route file
