+++ src/components/settings/EditNameDialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./EditNameDialog";
