+++ src/components/settings/EditResidenceDialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./EditResidenceDialog";
