+++ src/components/settings/SettingsGroup/index.ts
@@ -1,0 +1,1 @@
+export * from "./SettingsGroup";
