+++ src/components/settings/SettingsItem/index.ts
@@ -1,0 +1,1 @@
+export * from "./SettingsItem";
