+++ src/components/settings/index.ts
@@ -1,0 +1,6 @@
+export * from "./DeleteAccountDialog";
+export * from "./EditBirthplaceDialog";
+export * from "./EditNameDialog";
+export * from "./EditResidenceDialog";
+export * from "./SettingsGroup";
+export * from "./SettingsItem";
